refactor: generic stats component, gets stats data from outside
todo: Character가 뛰고 있는지 확인하는 부분이 적에게도 적용되려면 공유되는 요소가 있어야 함
